!!NON-WORKING!! idea: Idea for unit support
This is an idea for unit support in the engine. It is currently
non-working.
Idea:
- Create a units.txt file that contains definitions for how to convert
  between units.
- Create a "compiler" that generates zig functions from that file.
- Import that file in the engine and dynamically call functions from the
  generated file based on the lhs and rhs of conversion ("in" keyword)
  operations.

This commit contains a start, but it does not currently work (doesn't
even compile). It could be, that you can't just dynamically call
functions from structs like that, which would make this idea completely
useless.