feat: VERY basic implementation of units
This adds a VERY basic implementation of unit support for the engine. It
expands on the previous idea commit, but has been slightly modified.
Instead of generating functions for each conversion, it now generates
one big function, that takes in the number, the source unit and the
target unit. It then goes through a bunch of if statements to find out
which calculation needs to be used, before returning the result of that
calculation.
The engine then calls that conversion function with the unit of lhs and
rhs.