feat: Add support for unit
unit_gen.zig will now generates functionality for dealing with unit
prefixes ("c", "d", "g", ...). It handles them universally for all
units. It does this, by getting the index of the unit prefix for the
source unit and the target unit from a list (cronologically sorted;
biggest one on the right) and then either looping up or down the list.
While looping, it modifies the value that is later used for
conversion, multipling it by 10 when going down in the list, and
multiplying it by 0.1 when going up.

Then, after the number has been adjusted based on the prefix, we
continue with converting just as normal, according to the config from
units.txt.