feat: Only rewrite units.zig if units.txt was changed
unit_gen.zig will now only rewrite the units.zig file, if the units.txt
file has been changed. It detects this, by writing the last modified
timestamp into the first line of the output file. When generating again,
it compares that written time to the actual last modified timestamp of
the source file. If they are equal, the generation is cancelled, leaving
the output file as-is.