feat: Add basic gui application
This adds a foundation for builing a gui application, that uses the
calculator engine. Currently, when using "zig build run_gui", it shows
a window with the text "Hello World!". This can now be extended to
provide text fields for the user to type in and to show the results of
each calculation.

TODO: - Implement abstractions for the code currently in app/src/main.zig.
      - Use dark-mode. This is currently undocumented and after a first
	glance doesn't seem to be too trivial. I'm leaving it here as a
	todo for now, since it is not the most groundbreaking feature.