feat: Add widgets
This adds the concept of widgets to the gui application. Widgets are
more sophisticated controls that need custom logic and can't be rendered
via the paint system. One control that has been added so far is the text
area (self-explanatory). Widgets can be created using the
Application.makeWidget function, passing it the type and additional
options that may be used by the widget for customization.