feat: Calculate text field equation when pressing enter
The equation typed into the input field in the currently focused row
will now actually get calculated when pressing enter. The result of the
calculation is then written into the result text field on the right.

If you press Ctrl + Enter, you will not get send to a new row. The
current row will be calculated, but you'll stay in that row.