fix: Get rid of memory leaks by using arena allocators
The engine now uses two arena allocators for allocations. The first one
is a lasting allocator and is used for function and variable
declarations. The second one is a "temporary" allocator. It is created
before parsing and destroyed after the evaluation has finished. Being
arena allocators, we now don't have to worry about keeping track of
every allocated object. Since they aren't needed after calculation
anyways, we can just get rid of them by destroying the arena allocator.

The lasting allocator functions similarily, but it is created once at
the start and then destroyed at the end of the calculator lifecycle.
When exactly these events happen is decided by the user of the engine.