fix: Fix bug when using functions in function equations
Previously when using functions in function equations with a variable as
a parameter, you could only do the calculation once, but on every
further use, it would give you the same result as the first use.
This was because we were not deep copying the function equation and
thus expandVariables() modified the function equation, changing the
variable reference to a static operator. This has now been fixed by
actually deep copying the function equation before passing it to
evaluate().

FIXME: Rethink whether we actually need a deep copy. With the current
system (evaluate() and child functions modifing the ast) I don't think
it is easily avoidable without a deep copy, but maybe there's a better
way where we don't modify the ast at all.