fix: Correctly convert between unit prefixes
This commit changes two things: Firstly, the unit_prefixes array now has
a bunch of spacers ("__") to bridge the gap between e.g. milli (10^-3)
and micro (10^-6) while keeping the algorithm the same.
Secondly, the index for a unit without a prefix is now evaluated at
compile time, rather than being hardcoded and needing to be changed
everytime something changes in the array. At compile time, we just loop
over the array, looking for a value with "none". If there that can't be
found, a compile error is thrown.