feat: Re-implement the parser using the tokens from tokenizer
This re-implements the parser using the tokens generated from the
tokenizer. It generates the same AST as the old parser (the engine has
not been changed) but is way more compact and simple.
It currently supports everything except declarations and undeclarations
and it doesn't check whether functions or variables actually exist.