feat: Replace parser with tokenizer + parser
This replaces the previous parser with a tokenizer and a parser that uses the generated tokens. This makes it easier to add more features and properly handling whitespace in every situation. 
The parser now also works slightly different. It is now in a struct which makes it easier to recursively parse sub-groups.