feat: Add support for number modifiers (and add factorial "!")
This adds support for number modifiers, which currently include
factorial and more in the future. This also changed the way calculations
work, by creating an apply() function in the ast node, that takes in an
operator and another operand changes self's value depending on the
operator. This makes it easier do deal with different things such as
unit conversions or taking the modifier into consideration. It also
hides these details from the main code in engine.zig.

TODO: Make factorial work with groups and variables.