feat: Add equality checks
If you now put an equal sign ("=") at the top level of an equation (not
in a group / function parameter / ... that is), the engine will
calculate the left and right side separately and then compare the
result. After that, it will either return true or false, depending on
whether the two sides are equal.