fix: Correctly distinguish between the sign and the operator
This fixes a bug in the distinction between the sign and the operator
when encountering a + or a -. We now check if the last type was a
number, and if so, it is the operator, otherwise it is the sign. If it
is the sign, we save it to then later be added to the number.

This makes it now also possible to put the sign before a group, function
call or variable reference.