feat: Add support for utf8 characters ("°")
This adds support for utf8 characters, currently used to allow the
degrees symbol ("°"). If the tokenizer can't match the character by it's
ascii value, it will check whether it is the start of a unicode byte
sequence and if so, check if it is allowed and append it to the token
text. In the units.txt file, we can now use the zig unicode literal
("\u{....}") to have it correctly check against these characters.